refactor: move SidebarLink to its own file
